maintain music player between tab clicks
