+++ client/src/components/ProtectedRoutes.js
@@ -1,13 +1,35 @@
-import React  from "react";
+import React, { useState }  from "react";
 import { Outlet } from "react-router";
+
 import Login from "../pages/Login";
+import MusicPlayer from "../components/MusicPlayer";
 
 import { useUserContext } from "../UserContext";
 
 const ProtectedRoutes = () => {
     const { user } = useUserContext();
+    const [showPlayer, setShowPlayer] = useState(false);
+    const [currentTrack, setCurrentTrack] = useState({
+        id: "",
+        name: ""
+    });
 
-    return user?.username ? <Outlet/> : <Login/>;
+    const onItemClick = (id, name) => {
+        setCurrentTrack({ id, name });
+        setShowPlayer(true);
+    };
+
+    return user?.username ?
+        <>
+            <Outlet context={onItemClick}/>
+            {
+                showPlayer ?
+                    <MusicPlayer track={currentTrack.id} name={currentTrack.name} autoplay={true}/>
+                    :
+                    <></>
+            }
+        </>
+        : <Login/>;
 };
 
 export default ProtectedRoutes;
